renamed to RenderLabel
