+++ app/cdap/components/FooterPanel/Components/TableMetaInfoTab/index.tsx
@@ -15,7 +15,7 @@
  */
 
 import TabWrapper from 'components/FooterPanel/Components/common/TabWrapper';
-import TypographyLabel from 'components/FooterPanel/Components/common/TypographyLabel';
+import TypographyLabel from 'components/FooterPanel/Components/common/RenderLabel';
 import { PREFIX } from 'components/FooterPanel/constants';
 import T from 'i18n-react';
 import React from 'react';
