+++ app/cdap/components/FooterPanel/Components/common/RenderLabel/index.tsx
@@ -1,6 +1,6 @@
 import { Typography } from '@material-ui/core';
 import { grey } from '@material-ui/core/colors';
-import { ITypographyLabelProps } from 'components/FooterPanel/Components/common/TypographyLabel/types';
+import { IRenderLabelProps } from 'components/FooterPanel/Components/common/RenderLabel/types';
 import React from 'react';
 import styled from 'styled-components';
 
@@ -29,7 +29,7 @@ const OutlinedLabel = styled(Typography)`
  * @param type: simple or outlined, 2 variants of Typography Label
  * @returns TypographyLabel with appropriate variations according to props
  */
-export default function({ children, type }: ITypographyLabelProps) {
+export default function({ children, type }: IRenderLabelProps) {
   return (
     <>
       {type === 'simple' && (
